Add files via upload
The dataset used in this study consists of 658 petitions collected from Taiwan’s official government petition platform, Join(https://join.gov.tw). Each petition is assigned one or more of 12 predefined categories (labels). The petitions span the period from September 2015 to the end of January 2024. For the experiments, only petitions that had clearly defined categories on the website were included in the dataset.